safari-534.53-branch/LayoutTests/fast/js/finally-codegen-failure-expected.txt

Test that finally behaviour is correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f() is true
PASS (function () { var a = true; try { return a; } finally { a = false; }})() is true
PASS (function () { var a = 'PASS'; try { throw a; } finally { a = 'FAIL'; }})() threw exception PASS.
PASS successfullyParsed is true

TEST COMPLETE
